feat: Entity Highlighter updates (#112)
* Removed some unused CSS styling.
Added in Min/Max on SettingsTypes.Range
Added in stylizing for when SettingsTypse.Range hits a min or max.
Added in keyboard shortcuts to allow for setting up entity highlighting.
Added in handling for invisible blank entities.
Added in a priority button for adding/remove entities to the priority list.
Reorganized settings for plugin.

* Removed some unused CSS styling.
Added in Min/Max on SettingsTypes.Range
Added in stylizing for when SettingsTypse.Range hits a min or max.
Added in keyboard shortcuts to allow for setting up entity highlighting.
Added in handling for invisible blank entities.
Added in a priority button for adding/remove entities to the priority list.
Reorganized settings for plugin.

* Removed some unused CSS styling.
Added in Min/Max on SettingsTypes.Range
Added in stylizing for when SettingsTypse.Range hits a min or max.
Added in keyboard shortcuts to allow for setting up entity highlighting.
Added in handling for invisible blank entities.
Added in a priority button for adding/remove entities to the priority list.
Reorganized settings for plugin.
Fixed a function name typo.

* Delete package-lock.json

---------

Co-authored-by: KKonaOG <38236274+KKonaOG@users.noreply.github.com>